Move entity autocomplete to search module
