+++ apps/web/modules/components/entity/autocomplete/entity-text-autocomplete.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { useEffect, useRef } from 'react';
 import { Text } from '~/modules/design-system/text';
-import { useAutocomplete } from '~/modules/entity/autocomplete';
+import { useAutocomplete } from '~/modules/search';
 import { useSpaces } from '~/modules/spaces/use-spaces';
 import { Entity } from '~/modules/types';
 import { ResultContent, ResultsList } from './results-list';
