+++ apps/web/modules/search/index.ts
@@ -1,0 +1,1 @@
+export { useAutocomplete } from './autocomplete';
